(C) Sum Two Lowest Nums Fix for long integer

+++ c/src/katas/sum_two_lowest_nums/solution.c
@@ -2,8 +2,8 @@
 #include <stddef.h>
 
 long sum_two_smallest_numbers(size_t n, const int numbers[n]) {
-  int n1 = INT_MAX;
-  int n2 = INT_MAX;
+  long n1 = LONG_MAX;
+  long n2 = LONG_MAX;
 
   for (size_t i = 0; i < n; i++) {
     if (numbers[i] < n1) {
